+++ src/main/java/org/betonquest/betonquest/quest/event/weather/WeatherEventFactory.java
@@ -7,14 +7,13 @@ import org.betonquest.betonquest.api.common.function.Selector;
 import org.betonquest.betonquest.api.common.function.Selectors;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.quest.event.Event;
-import org.betonquest.betonquest.api.quest.event.EventFactory;
+import org.betonquest.betonquest.api.quest.event.HybridEvent;
+import org.betonquest.betonquest.api.quest.event.HybridEventFactory;
 import org.betonquest.betonquest.api.quest.event.StaticEvent;
-import org.betonquest.betonquest.api.quest.event.StaticEventFactory;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.quest.event.DoNothingStaticEvent;
-import org.betonquest.betonquest.quest.event.NullStaticEventAdapter;
 import org.betonquest.betonquest.quest.event.OnlineProfileRequiredEvent;
-import org.betonquest.betonquest.quest.event.PrimaryServerThreadEvent;
+import org.betonquest.betonquest.quest.event.PrimaryServerThreadHybridEvent;
 import org.bukkit.Server;
 import org.bukkit.World;
 import org.bukkit.entity.Player;
@@ -27,7 +26,7 @@ import java.util.Locale;
 /**
  * Factory to create weather events from {@link Instruction}s.
  */
-public class WeatherEventFactory implements EventFactory, StaticEventFactory {
+public class WeatherEventFactory implements HybridEventFactory {
     /**
      * Logger factory to create a logger for events.
      */
@@ -64,14 +63,17 @@ public class WeatherEventFactory implements EventFactory, StaticEventFactory {
     }
 
     @Override
-    public Event parseEvent(final Instruction instruction) throws InstructionParseException {
+    public HybridEvent parseHybridEvent(final Instruction instruction) throws InstructionParseException {
         final Weather weather = parseWeather(instruction.next());
         final Selector<World> worldSelector = parseWorld(instruction.getOptional("world"));
         final VariableNumber duration = instruction.getVarNum(instruction.getOptional("duration", "0"));
-        return new PrimaryServerThreadEvent(
-                new OnlineProfileRequiredEvent(
-                        loggerFactory.create(WeatherEvent.class), new WeatherEvent(weather, worldSelector, duration), instruction.getPackage()),
+        return new PrimaryServerThreadHybridEvent(new WeatherEvent(weather, worldSelector, duration),
                 server, scheduler, plugin);
+    }
+
+    @Override
+    public Event parseEvent(final Instruction instruction) throws InstructionParseException {
+        return new OnlineProfileRequiredEvent(loggerFactory.create(WeatherEvent.class), parseHybridEvent(instruction), instruction.getPackage());
     }
 
     @Override
@@ -79,7 +81,7 @@ public class WeatherEventFactory implements EventFactory, StaticEventFactory {
         if (instruction.copy().getOptional("world") == null) {
             return new DoNothingStaticEvent();
         } else {
-            return new NullStaticEventAdapter(parseEvent(instruction));
+            return parseHybridEvent(instruction);
         }
     }